fix(playerbot): Move InstanceProgress to interface
- Moved InstanceProgress struct to IInstanceCoordination.h
- Added GameTime.h include for GameTime::GetGameTimeMS()
- Removed duplicate from InstanceCoordination.h
